requirements.txt

mlflow==2.10.0
python-dotenv==1.1.0

flask==3.0.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.5.2
psycopg2-binary==2.9.9
tcgm==0.1.4
pickle5==0.0.12
gunicorn==21.2.0
```

2. **Create `Procfile`** (no extension):
```
web: gunicorn api_service:app